# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HRPay API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, on_leave, terminated]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete (terminate) an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- DEPARTMENTS ---
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- PAYROLL ---
...
  /leave/requests:
    get:
      operationId: listLeaveRequests
      tags: [leave]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, cancelled]
        - name: type
          in: query
          schema:
            type: string
            enum: [vacation, sick, personal, maternity, paternity, unpaid]
      responses:
        "200":
          description: List of leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leave]
      summary: Create a leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveRequestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave/requests/{id}/approve:
    post:
      operationId: approveLeaveRequest
      tags: [leave]
      summary: Approve a leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave/requests/{id}/reject:
    post:
      operationId: rejectLeaveRequest
      tags: [leave]
      summary: Reject a leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave/balances:
    get:
      operationId: listLeaveBalances
      tags: [leave]
      summary: Get leave balances
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leave balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveBalance"

  # --- BENEFITS ---
  /benefits/plans:
    get:
      operationId: listBenefitPlans
      tags: [benefits]
      summary: List benefit plans
      responses:
        "200":
          description: List of benefit plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BenefitPlan"
    post:
      operationId: createBenefitPlan
      tags: [benefits]
      summary: Create a benefit plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBenefitPlanBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenefitPlan"
  /benefits/enrollments:
    get:
      operationId: listBenefitEnrollments
      tags: [benefits]
      summary: List benefit enrollments
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: planId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BenefitEnrollment"
    post:
      operationId: createBenefitEnrollment
      tags: [benefits]
      summary: Enroll employee in benefit plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBenefitEnrollmentBody"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenefitEnrollment"

  # --- ANALYTICS (Wow endpoints) ---
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, contractor, intern]
        status:
          type: string
          enum: [active, inactive, on_leave, terminated]
        startDate:
          type: string
          format: date
        salary:
          type: number
        salaryType:
          type: string
          enum: [annual, hourly]
        managerId:
          type: integer
        managerName:
          type: string
        avatarUrl:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        taxId:
          type: string
        bankAccount:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - position
        - employmentType
        - status
        - startDate
    EmployeeListResponse:
      type: object
      properties:
        employees:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - employees
        - total
        - page
        - limit
    CreateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        departmentId:
          type: integer
        employmentType:
          type: string
          enum: [full_time, part_time, contractor, intern]
        startDate:
          type: string
          format: date
        salary:
          type: number
        salaryType:
          type: string
          enum: [annual, hourly]
        managerId:
          type: integer
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
      required:
        - firstName
        - lastName
        - email
        - position
        - employmentType
        - startDate
    UpdateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        departmentId:
          type: integer
        employmentType:
          type: string
          enum: [full_time, part_time, contractor, intern]
        status:
          type: string
          enum: [active, inactive, on_leave, terminated]
        salary:
          type: number
        salaryType:
          type: string
          enum: [annual, hourly]
        managerId:
          type: integer
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        headCount:
          type: integer
        managerId:
          type: integer
        managerName:
          type: string
        budget:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        managerId:
          type: integer
        budget:
          type: number
      required:
        - name
...
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        type:
          type: string
          enum: [vacation, sick, personal, maternity, paternity, unpaid]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        days:
          type: number
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        reviewedBy:
          type: string
        reviewedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - type
        - startDate
        - endDate
        - status
    CreateLeaveRequestBody:
      type: object
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [vacation, sick, personal, maternity, paternity, unpaid]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
      required:
        - employeeId
        - type
        - startDate
        - endDate
    LeaveBalance:
      type: object
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        type:
          type: string
          enum: [vacation, sick, personal]
        allocated:
          type: number
        used:
          type: number
        pending:
          type: number
        remaining:
          type: number
        year:
          type: integer
      required:
        - employeeId
        - type
        - allocated
        - used
        - remaining
    BenefitPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [health, dental, vision, life, disability, retirement, other]
        provider:
          type: string
        description:
          type: string
        employeeCost:
          type: number
        employerCost:
          type: number
        enrolledCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - isActive
    CreateBenefitPlanBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [health, dental, vision, life, disability, retirement, other]
        provider:
          type: string
        description:
          type: string
        employeeCost:
          type: number
        employerCost:
          type: number
      required:
        - name
        - type
    BenefitEnrollment:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        planId:
          type: integer
        planName:
          type: string
        enrolledAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - id
        - employeeId
        - planId
        - isActive
    CreateBenefitEnrollmentBody:
      type: object
      properties:
        employeeId:
          type: integer
        planId:
          type: integer
      required:
        - employeeId
        - planId
...