# lib/api-zod/src/generated/types/createBenefitEnrollmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HRPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBenefitEnrollmentBody {
  employeeId: number;
  planId: number;
}