# lib/api-zod/src/generated/types/createBenefitPlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HRPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBenefitPlanBodyType } from "./createBenefitPlanBodyType";

export interface CreateBenefitPlanBody {
  name: string;
  type: CreateBenefitPlanBodyType;
  provider?: string;
  description?: string;
  employeeCost?: number;
  employerCost?: number;
}