# lib/api-zod/src/generated/types/updateEmployeeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HRPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEmployeeBodyEmploymentType } from "./updateEmployeeBodyEmploymentType";
import type { UpdateEmployeeBodySalaryType } from "./updateEmployeeBodySalaryType";
import type { UpdateEmployeeBodyStatus } from "./updateEmployeeBodyStatus";

export interface UpdateEmployeeBody {
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  position?: string;
  departmentId?: number;
  employmentType?: UpdateEmployeeBodyEmploymentType;
  status?: UpdateEmployeeBodyStatus;
  salary?: number;
  salaryType?: UpdateEmployeeBodySalaryType;
  managerId?: number;
  address?: string;
  city?: string;
  state?: string;
  country?: string;
}